Prepare for PHP8

+++ lib/routing/sfRoute.class.php
@@ -264,7 +264,7 @@ class sfRoute implements Serializable
 
   static private function generateCompareVarsByStrlen($a, $b)
   {
-    return strlen($a) < strlen($b);
+    return (int) (strlen($a) < strlen($b));
   }
 
   /**
